Change FrameNumber from u16 to u32 for now
A premature optimisation that was starting to get annoying with all
the Wrapping safe math. May change back later.

+++ src/lib.rs
@@ -161,7 +161,7 @@ pub mod prelude {
     pub use crate::DespawnMarker;
     pub use crate::TimewarpConfig;
     pub use crate::TimewarpPlugin;
-    pub type FrameNumber = u16;
+    pub type FrameNumber = u32;
     pub use crate::TimewarpSet;
 }
